third_party/WebKit/LayoutTests/http/tests/security/XFrameOptions/x-frame-options-parent-same-origin-deny.html

<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.dumpChildFramesAsText();
        testRunner.waitUntilDone();
    }

    function checkIfDone() {
        try {
            var url = document.querySelector('iframe').contentWindow.location.href;
            console.log("FAIL: Could read contentWindow.location.href");
        } catch (e) {
            console.log("PASS: Access to contentWindow.location.href threw an exception.");
        }

        testRunner.notifyDone();
    }
</script>

<p>There should be no content in the iframe below</p>
<iframe style="width:500px; height:500px" src="http://localhost:8000/security/XFrameOptions/resources/x-frame-options-parent-same-origin-allow.cgi" onload="checkIfDone()"></iframe>
