third_party/WebKit/LayoutTests/http/tests/security/XFrameOptions/x-frame-options-parent-same-origin-deny-expected.txt

CONSOLE ERROR: Refused to display 'http://localhost:8000/security/XFrameOptions/resources/x-frame-options-parent-same-origin-allow.cgi' in a frame because it set 'X-Frame-Options' to 'sameorigin'.
CONSOLE MESSAGE: line 13: PASS: Access to contentWindow.location.href threw an exception.
There should be no content in the iframe below



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
